Refactor: Reuse pruning type definitions and simplify HTML/Text pruners
